code filter function for categories list, refactor

+++ src/main/java/com/viettel/shopme/admin/user/controller/UserController.java
@@ -48,7 +48,7 @@ public class UserController {
         Page<User> page = service.listByPage(pageNum, sortField, sortDir, keyword);
         List<User> listUsers = page.getContent();
 
-        long startCount = (pageNum - 1) * UserService.USER_PER_PAGE + 1;
+        long startCount = (long) (pageNum - 1) * UserService.USER_PER_PAGE + 1;
         long endCount = startCount + UserService.USER_PER_PAGE -1;
         if (endCount > page.getTotalElements()) {
             endCount = page.getTotalElements();
